Fix an error for `Layout/AccessModifierIndentation`
This PR fixes the following infinite loop error for `Layout/AccessModifierIndentation`
when the access modifier is on the same line as the class definition.

```console
$ echo 'class A; private; def foo; end; end' | bundle exec rubocop --stdin dummy.rb -a --only Layout/AccessModifierIndentation
(snip)

Infinite loop detected in /Users/koic/src/github.com/rubocop/rubocop/dummy.rb and caused by Layout/AccessModifierIndentation
Hint: Please update to the latest RuboCop version if not already in use, and report a bug if the issue still occurs on this version.
Please check the latest version at https://rubygems.org/gems/rubocop.
/Users/koic/src/github.com/rubocop/rubocop/lib/rubocop/runner.rb:322:in 'block in RuboCop::Runner#iterate_until_no_changes'
<internal:kernel>:191:in 'Kernel#loop'
```

In cases like this, where they are on the same line, the concept of indentation doesn't apply, so it can be ignored.