Set `EnforcedShorthandSyntax: either` by default for `Style/HashSyntax`
Follow up https://github.com/rubocop/rubocop/pull/10351#issuecomment-1008235461

As the current trend suggests, `EnforcedShorthandSyntax: either` seems to be a reasonable choice.
Additionally, as a note from the developer, `either` was added later as an option for `EnforcedShorthandSyntax`.
Had it been noticed from the beginning, it would likely have been the default.
`EnforcedShorthandSyntax: always` is still maintained as an option, and users who prefer it will need to switch the setting.
Rather than switching from `EnforcedShorthandSyntax: always` to the extreme of `EnforcedShorthandSyntax: never`,
`EnforcedShorthandSyntax: either` leaves the choice to the user, making it a more balanced decision in that regard.