[Fix #12597] Update `Style/SingleLineDoEndBlock` to not register an offense if it will introduce a conflicting `Layout/RedundantLineBreak` offense.
If `Layout/RedundantLineBreak` is enabled and configured with `InspectBlocks: true`, then blocks will be rewritten to be on a single line if safe to do so. However, if `Style/SingleLineDoEndBlock` is also enabled, those single line blocks will then be corrected back to a multiline block, thus beginning an infinite loop.

In this case, no longer register a `Style/SingleLineDoEndBlock` offense if it can be a single line. Blocks that cannot be expressed on a single line will still be registered.