Fix RedundantLineContinuation false positive on method def argument
This change fixes a false positive for `Style/RedundantLineContinuation`
when a method definition (which always returns a symbol) is used as an
argument for a method call.

In such a situation, a backslash to explicitly indicate the line
continuation is needed (between the name of the method being called and
the method definition that is being used as an argument). Prior to this
change, however, RuboCop would incorrectly complain that the backslash
is not needed (and autocorrection would remove the needed backslash).