Fix infinite loop error for `Style/IfWithSemicolon`
This PR fixes infinite loop error for `Style/IfWithSemicolon`
when using nested if/;/end in if body.

```console
$ cat example.rb
if cond1; foo + if cond2; bar
  else
  end
end
$ bundle exec rubocop -a -d
(snip)
Parser::Source::TreeRewriter detected clobbering
/Users/koic/.rbenv/versions/3.4-dev/lib/ruby/gems/3.4.0+0/gems/parser-3.3.6.0/lib/parser/source/tree_rewriter.rb:427:
in 'Parser::Source::TreeRewriter#trigger_policy'
/Users/koic/.rbenv/versions/3.4-dev/lib/ruby/gems/3.4.0+0/gems/parser-3.3.6.0/lib/parser/source/tree_rewriter.rb:414:
in 'Parser::Source::TreeRewriter#enforce_policy'
/Users/koic/.rbenv/versions/3.4-dev/lib/ruby/gems/3.4.0+0/gems/parser-3.3.6.0/lib/parser/source/tree_rewriter/action.rb:234:
in 'Method#call'
```

The autocorrection for the nested conventional case is handled as a repeated autocorrection,
so there is no impact on the user experience.