[Fix #13320] Fix incorrect autocorrect when `Layout/LineContinuationLeadingSpace` and `Style/StringLiterals` autocorrects in the same pass
When both cops are activated and run in the same iteration of the correction loop, `Style/StringLiterals` undoes the moving of spaces that `Layout/LineContinuationLeadingSpace` performs. This is because `StringLiterals` takes the original string content and transforms it, rather than just modifying the delimiters, in order to handle escaping for quotes within the string.

By marking `Style/StringLiterals` as incompatible with `Layout/LineContinuationLeadingSpace`, it allows the correction to be deferred to the next inspection loop, thereby avoiding the problem.