[Fix #13284] Add new `target_gem_version` API
This allows cops that require a gem to handle multiple different versions
of a gem at the same time.
Previously the only piece of information you had was that the gem was present.

For example in rack 3.1, new status code aliases were
added that should not be handled on earlier versions.

Since this only looks at the lockfile, default gems are not considered.