[Fix #13299] Fix `Style/BlockDelimiters` to always accept braces when an operator method argument is chained
Update `Style/BlockDelimiters` to not register an offense when braces are necessary to avoid changing precedence. Previously when an operator method was encountered, it was automatically evaluated (ie. not marked as ignored), but this can result in block delimiters being changed that affects block binding.

Now operator methods are evaluated for nodes to ignore, other than if it only has a single argument which is of block type.

As well, I added the missing `expect_corrections` tests for offenses without it.