[Fix #7958] Properly tag unsafe autocorrections with `rubocop:todo` when using `--autocorrect --disable-uncorrectable`
Previously, if a cop was correctable but not corrected due to being unsafe, `--autocorrect --disable-uncorrectable` would report the offense but not tag it with `rubocop:todo`. This is now updated, so that unsafe offenses are tagged as well.

In order to do so, we have to treat all cops as "autocorrect_cops" in `Team#investigate`, since adding the directive is also a correction; otherwise, the offense report will mark the offense as `[Todo]` but not actually apply the correction.