[Fix #13556] Fix false positives for `Style/FileNull`
Fixes #13556.

This PR fixes false positives for `Style/FileNull` when using `'nul'` string.

This PR ensures only files that use the string `'/dev/null'` are targeted for detection.
This is because the string `'nul'` is not limited to the null device.
This behavior results in false negatives when the `'/dev/null'` string is not used,
but it is a trade-off to avoid false positives.