Add support for Ruby LSP as a built-in add-on
## Summary

This PR adds support for the Ruby LSP Add-on:
https://shopify.github.io/ruby-lsp/add-ons.html

Ruby LSP is widely used among Ruby developers.

This Add-on resolves the current dependency of Ruby LSP on RuboCop's internal APIs
by allowing RuboCop itself to provide the features related to Ruby LSP,
marking the start of potentially removing these dependencies in the future.

It may also be possible to integrate the Standard Ruby's LSP implementation into
this RuboCop's LSP implementation at a later time.
A future is envisioned in which Ruby LSP, Standard Ruby, and RuboCop
don't each need to implement their own RuboCop runner for
LSP but instead unify that functionality.

An option is available so it can work with both RuboCop's built-in LSP and Ruby LSP.
RuboCop's built-in LSP will continue to serve as a lightweight LSP for users who only need RuboCop.

## References

- https://github.com/standardrb/standard/pull/630
- https://github.com/standardrb/standard/wiki/IDE:-vscode#using-ruby-lsp

## NOTE

This PR is split into multiple commits around three core PRs related to Standard Ruby.
Please don't squash these commits, as the separation is intentional.

Documentation will likely be covered later in conjunction with Ruby LSP.