Add support for Code Actions under Ruby LSP
Follow up https://github.com/standardrb/standard/pull/636.

Because standardrb/standard#636 alone does not behave as expected in VS Code,
this includes the content of the following commits related to the changes in
standardrb/standard#636.

- https://github.com/standardrb/standard/commit/ab18144a677f76d50d37aaeae2f4b75a3f7422d1
- https://github.com/standardrb/standard/commit/edd199b6b594560a3ed81a9687a8bb394cf2fd24
- https://github.com/standardrb/standard/commit/4240fc5878fd5f62ce5869c6916571b780b6fc1d
- https://github.com/standardrb/standard/commit/2fcca8d4d060b1cc3daf7dc86227b397b86ab6d3
- https://github.com/standardrb/standard/commit/84ee9f4f203eef9e35f93ff033cebcf64c007959

This ensures that the features related to standardrb/standard#636 will work,
so it is being submitted as an independent commit.