[Fix #13677] Fix false negative on `Style/RedundantLineContinuation` when the continuation is followed by a percent array
Previously, `start_with_arithmetic_operator?` used a regexp to check if the following line starts with an arithmetic operator (`+`, `-`, `*`, `/`, '**' or `%`). However, a percent array (eg. `%i[foo bar]`) also starts with a `%`, and as such, was being marked as a required continuation.

This has been corrected now to check against the first token of the line rather than the first character.