Append timestamp to generated changelog files to avoid clobbering
When the first 50 characters of a commit message are the same in multiple commits, `rake changelog:*` will generate the same filename for each of the commits, resulting in file clobbering and missed changelog entries.

By appending the current timestamp to the end of the filename, we should avoid this issue.

NB: A byproduct of this is that running `rake changelog:*` again on the same commit will generate an additional changelog entry.