Fix `ExpectOffense` when a source line contains an escaped caret
Previously in f450f62, In order to be able to handle lines that start with a caret, I updated `expect_no_offense` to escape carets, since there are no annotations expected with that method, in order to allow lines that start with a caret. Then in `AnnotatedSource.parse`, the caret was unescaped so that the source actually matches what was given.

However, this caused tests to fail in rubocop-performance, which contains escaped carets inside a regexp.

I've updated again to not add any escaping, but rather not parse the source given to `expect_no_offenses` for annotations, as there should not be any in the first place.