Add new `Lint/CopDirectiveSyntax` cop
When disabling specific cops within source code, incorrect
formatting of directive comments can lead to unintended
behavior.

For instance:

  # rubocop:disable Naming/MethodName Style/EmptyMethod
  #                                  ^ comma is missing

In this example, `Style/EmptyMethod` is not disabled as intended,
because the directive lacks the required comma.

This cop ensures that directive comments are strictly formatted
to avoid such issues.

Additionally, we have officially defined the `--` prefix as the
standard way to add comments to cop directives and updated the
documentation:

  # rubocop:disable Naming/MethodName -- This is a comment.