[Fix #13745] Add new `InternalAffairs/NodePatternGroups` cop
The cop detects cases where node patterns contain node types that can be replaced with a node group (ie. `{send csend}` can be `call`).

It uses `RuboCop::AST::NodePattern::Parser::WithMeta` to parse node patterns found in `def_node_matcher` or `def_node_search` calls, with a custom processor and AST walker to detect the exact nodes with the node pattern that are replaceable. The metadata generated while walking the Node Pattern AST is then used to highlight offenses in the node pattern string, and make corrections.

Correction is extensively tested, and I have run the cop, including corrections, on all the RuboCop extensions, and opened a number of PRs to fix issues identified via this cop.