[Fix #13750] Fix false positives for `Style/RedundantSelfAssignment`
Fixes #13750.

This PR fixes false positives for `Style/RedundantSelfAssignment`
when assigning to attribute of `self`.

As mentioned in the feedback on #13750, `self.foo = foo.concat(ary)` is an assignment through a method call,
which differs from a variable assignment like `foo = foo.concat(ary)`.

Although the original example suggests replacing `self.foo = foo.concat(ary)` with `self.foo += arg`,
the autocorrect implementation didn't account for this. Furthermore, methods other than `concat`,
such as `delete_if` or `unshift`, cannot be resolved by replacing them with `+=`.
As a result, providing a simple autocorrect solution is not feasible.
Furthermore, it is doubtful whether such a change can be regarded as an improvement to good style.

Given this, better approach would likely be to avoid detecting assignments to `self.foo` in such cases.