Make the language_server-protocol dependency version stricter
The `language_server-protocol` gem is a library that adheres to the LSP specification.
Since `X.Y.Z` represents the LSP version:
https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/

Only the fourth value can be updated without affecting compatibility with the LSP specification:
https://github.com/mtsmfm/language_server-protocol-ruby?tab=readme-ov-file#versioning

This PR strictly defines the runtime dependency version of the `language_server-protocol` gem
in advance to prevent unintended incompatibilities.