Introduce EnforcedStyleForMultiline "diff_comma"
[Fix #9935]

A new "diff_comma" enforced style for `Style::TrailingCommaInArrayLiteral`
and `Style::TrailingCommaInHashLiteral` that omits the trailing comma unless
the final item of the hash/array literal is at the end of line.

This style behaves similarly to "consistent_comma" with the exception of this case:

```ruby
 # consistent_comma behavior
 ## bad
 a = { foo: 1, bar: 2,
       baz: 3, qux: 4 }

 ## good
 a = { foo: 1, bar: 2,
       baz: 3, qux: 4, }

 # diff_comma behavior
 ## bad
 a = { foo: 1, bar: 2,
       baz: 3, qux: 4, }

 ## good
 a = { foo: 1, bar: 2,
       baz: 3, qux: 4 }
```

In #9935 it was discussed whether this should be the behavior of
"consistent_comma" but some concerns were raised about whether this
was expected behavior or not, or whether the change would be
disruptive.

This style convention results in "clean diffs" where adding an item to
the hash/array literal results in a diff like:

```patch
 {
   foo: 123,
+  bar: 456,
 }
```

without the downside of unnecessary "consistent_comma" commas like
this when placing a closing bracket on the same line as the final
argument:

```
 { foo: 123,
   bar: 456, }
           ^ like this one
```