Make `SuggestExtensions` aware of plugin extensions
This PR updates `SuggestExtensions` to properly handle plugin extensions.

For example, when changing the plugin specification for `rubocop-rake` from `require` to `plugins`:

```diff
 inherit_from: .rubocop_todo.yml

-require:
-  - rubocop/cop/internal_affairs
+plugins:
+  - rubocop-internal_affairs
   - rubocop-rake
+
+require:
   - rubocop-rspec

 AllCops:
```

Currently, `rubocop-rake` is unexpectedly displayed as part of `SuggestExtensions`:

```console
$ bundle exec rubocop
(snip)

The following RuboCop extension libraries are installed but not loaded in config:
  * rubocop-rake

You can opt out of this message by adding the following to your config
(see https://docs.rubocop.org/rubocop/extensions.html#extension-suggestions for more options):
  AllCops:
    SuggestExtensions: false
```

Ideally, when a plugin is specified under `plugins`, it should not be displayed by `SuggestExtensions`,
just like when it is specified under `require`.

Since this is a follow-up to the unreleased #13792, there is no changelog entry.