Fix exclusion of relative paths in plugin's `AllCops: Exclude` as expected
When relative paths are used in a plugin's configuration file under `AllCops: Exclude`,
they must be converted to absolute paths.

Otherwise, the exclusion paths will not match the paths being inspected, and the files will not be excluded.

This issue was reported in https://github.com/rubocop/rubocop-rails/issues/1443.

This issue was previously reported in https://github.com/rubocop/rubocop-rails/pull/187
and is essentially a necessary process for all extension gems.

On a side note, converting to the plugin system allows us to fix such potential bugs across all extension gems in a unified manner.
Of course, there is also the risk of introducing bugs uniformly...