[Fix #13826] Fix false positives for regex cops when `Lint/MixedCaseRange` is enabled
The `RegexpRanges` class was added to work around a bug in regexp_parser with ocals in ranges.
But, it destructivly removes all expressions from sets because it doesn't dup them.
So, any cop that runs after that wants to look at sets will see nothing at all.

Today, the buggy versions are excluded in the gemspec, so the code can simply be removed.

I didn't add a new test because the code that would be tested is gone now.

It also fixes a false positive for `/[_A-z;&&[^G-f]]/`. This regex only contains a single range, not two.