[Fix #13876] Don't consider `pp` for `Lint/RedundantRequireStatement`
Only the `pp` method is available. Requiring it brings in a bunch of other stuff, like the `PP` constant or a few `pretty_` methods.

There's code to handle this but it only works when those methods are exclusivly used in the file it is required in, which
I don't think is a very realistic expectation.

All the other requires are truly redundant, this one looks out of place (especially for `Lint`)