Allow merging of configured arrays and non-arrays
Previously, merging would only be performed if the base value was an
`Array`. If the derived value was an `Array`, but the base value was
not, an overwrite would be performed.

For example, many cops specify a single `Reference` directly as a
`String`. However, as `Reference` uses `Array(value)` behind the scenes,
and accepts either a `String` or an `Array` of strings, a consumer may
decide to add additional references.

    # config/default.yml
    Security/JSONLoad:
      Reference: 'https://ruby-doc.org/stdlib-2.7.0/libdoc/json/rdoc/JSON.html#method-i-load'

    # .rubocop.yml
    Security/JSONLoad:
      Reference:
        - https://example.com#another-reason-to-use-json-parse-instead-of-load

Unfortunately, this would overwrite the original, even if configured to
merge..

By looking at both the base and derived values to check for the presence
of an `Array`, we can detect that a merge may be appropriate, and by
wrapping both values in `Array(...)`, we can allow the merge to succeed,
resulting in the intended config combination:

    Security/JSONLoad:
      Reference:
        - https://ruby-doc.org/stdlib-2.7.0/libdoc/json/rdoc/JSON.html#method-i-load
        - https://example.com#another-reason-to-use-json-parse-instead-of-load

It remains the responsibility of the config validation phase and
individual cops to ensure arrays are rejected where they are unexpected.

This change technically makes it possible to extend array configurations
by providing a bare string, although the intention is to support the
reverse (specifically for the `Reference` use case).