Make `CopsDocumentationGenerator` support plugin system for extensions
`CopsDocumentationGenerator` is a private API used primarily for generating documentation for
extension gems managed by RuboCop HQ.

With the plugin migration in #13792, the implementation was changed to use `RuboCop::ConfigLoader.inject_defaults!`
instead of `Inject`, as shown below:

```diff
-  RuboCop::Performance::Inject.defaults!
+  RuboCop::ConfigLoader.inject_defaults!("#{__dir__}/../config/default.yml")
```

However, switching to this API caused issues in https://github.com/rubocop/rubocop-rspec/pull/2047.
This PR addresses those issues in a centralized manner within `CopsDocumentationGenerator`.

As a result, the changes in extension gems will look like this:

```diff
-  RuboCop::ConfigLoader.inject_defaults!("#{__dir__}/../config/default.yml")
-
-  CopsDocumentationGenerator.new(departments: deps).call
+  CopsDocumentationGenerator.new(departments: deps, plugin_name: 'rubocop-performance').call
```

One of the benefits of this change is that it reduces the need to use the `RuboCop::ConfigLoader.inject_defaults!` API.