[Fix #12669] Remove unsafe autocorrection for `Lint/EmptyConditionalBody`
In #10862, autocorrection was added to remove branches that had no body. This was made unsafe in #10867, because deleting a branch introduces a behaviour change if any of the branch conditions had side effects.

As a `Lint` cop, it'd be better to just not correct when unsafe. The user can decide whether the branch can be removed, or an explicit `nil` can be added.