Fix false positives for `Style/RedundantCondition`
This PR fixes false positives for `Style/RedundantCondition`
when a variable or a constant is used.

This fixes false positives that were not addressed in #13919.
For example, logic that transforms `variable = 42` into `variable ? true : false`
should not be detected. Since it is fundamentally impossible to determine
what values a variable or constant might be assigned, they are always allowed.