Allow writing Steep annotation on the same line as `def`, `end`, `begin`, `class` and `module` in `Style/CommentedKeyword`
[Steep](https://github.com/soutaro/steep) is a type checker for Ruby.

Under certain circumstances, if the programmer writes `steep:ignore`, Steep will not do any type checking.

In particular, `steep:ignore` can be written on the same line as `def`, `end`, `begin`, `class` and `module`, but it is rejected by the `Style/CommentedKeyword` rules.

Another way to do this is to add options, but once an option is added, it is difficult to remove it, so the default rule written.

Also, it is unlikely that you would write `steep:ignore` if you were not using steep, so I defined it as a default rule.

refs: https://github.com/soutaro/steep/wiki/Release-Note-1.7#ignore-diagnostics-by-steepignore-comment