Fix a false negative for `Lint/ReturnInVoidContext` when returning out of a block
Should be block happen to be called, the return will break out of the block and attribute the return value to the method.

This has been that way since the cop was added and there were no tests for blocks.

The only expection to the rule is when returning from inside a lambda.

Additionally, I simplified the implementation. The cop looked for `sdef` but then immediatly after
required that the node be simply `def`.