[Fix #12360] Add new `Naming/PredicateMethod` cop to check that predicate methods end with `?` and non-predicate methods do not
In `Mode: conservative` (default), question marks are required when all return values are boolean, and forbidden when all are known to be non boolean (literals). Otherwise, methods are accepted with and without a question mark, in order to make fixing offenses manageable.

In `Mode: aggressive`, methods are not allowed to have a question mark if any return values are known to be non-boolean.