Update `Style/RedundantFormat` to register offenses when the only argument to `format` or `sprintf` is a constant
Since `format` and `sprintf` require the first argument to be a string (and otherwise raises `TypeError`), it is safe to assume that `format` with a single constant argument can be replaced with the constant itself, in the same way as a string argument.

If the constant is splatted, it will not register an offense.