Ignore permission errors when searching for gemspec files
I'm attempting to run rubocop (and other test tools) in an environment
constrained by sandbox-exec[1] to improve reproducibility. In this
sandboxed environment, tools only have access to a read-only copy of the
project dirctory, a fake writeable home directory, and a writeable temp
directory. When attempting to run rubocop in this setup, it fails to due
permission errors while attempting to search for gemspec files in parent
directories outside the sandbox. In this scenario I think it's
reasonable to treat hidden gemspecs the same as gemspeces that don't
exist.

[1] https://reverse.put.as/wp-content/uploads/2011/09/Apple-Sandbox-Guide-v1.0.pdf