[Fix #13617] Proper support for Ruby 3.4
Now that `rubocop-ast` uses `prism` for 3.4, all the syntax is supported
and it no longer needs to me marked experimental.

As you can seen, there's not much actually happening here, mainly
just doc updates. That is because this behaviour is entirely controlled
in `rubocop-ast` when `ParserEngine: default` is set.

This just bumps the `rubocop-ast` requirement to always get a version where that is the case.