Tweak the output of `rubocop -V` when `TargetRubyVersion` <= 3.2
Follow-up to https://github.com/rubocop/rubocop/pull/14025

This PR tweaks the output of `rubocop -V` when `TargetRubyVersion` <= 3.2.

Since Prism only parses Ruby 3.3 and above, it won't be used when `TargetRubyVersion` is lower than 3.2.
Therefore, there's no need to display it in the output of `rubocop -V`.

The use of Prism for `TargetRubyVersion` 3.3 and 3.4 depends on the specific case,
but the output is already supplemented by `parser_engine_text`, so the current display format is kept as is.