Introduce `References` to soft-replace `Reference`
Given references are an array of strings, it makes sense for the config to
be pluralized. As we're changing the name, we also take the opportunity
to add a warning if the config is a string instead of an array. Given
this is a new name, it shouldn't cause any warnings in any existing
codebases or plugins, but helps ensure that when consumers migrate to
the new name, they also migrate to an array.

Support for the legacy `Reference` key is maintained, including support
for providing a simple string instead of an array.