Add `EnforcedStyle: allow_single_line` as the default to `Style/ItBlockParameter`
This PR adds `EnforcedStyle: allow_single_line` as the default to `Style/ItBlockParameter`.

This adds the default configuration that follows `EnforcedStyle: allow_single_line (default)` of
`Style/NumberedParameters` cop:
https://docs.rubocop.org/rubocop/1.75/cops_style.html#stylenumberedparameters

In RuboCop, numbered block parameters are allowed only in single-line blocks.
It makes sense to apply the same default style to `it` block parameters as well.