Tweak autocorrect for `Style/LambdaCall`
This commit tweaks autocorrect for `Style/LambdaCall` when using `()` with no arguments:

```console
$ echo '42.()' | bundle exec rubocop --stdin test.rb --only Style/LambdaCall -a
Inspecting 1 file
C

Offenses:

test.rb:1:1: C: [Corrected] Style/LambdaCall: Prefer the use of 42.call() over 42.().
42.()
^^^^^

1 file inspected, 1 offense detected, 1 offense corrected
====================
42.call()
```

In practice, `call` is preferred over `call()`, and this can be adjusted with a minimal change
in the cop implementation.
This is only a minor adjustment to the message and autocorrection, so no changelog entry is included.