Layout/LeadingCommentSpace: allow splitting long inline RBS comment signatures
At Shopify we use [RBS inline comments with Sorbet](https://sorbet.org/docs/rbs-support).

We added a syntax to allow splitting long signatures accross multiple lines
like this:

```rb
\#: (
\#|   Integer,
\#|   String
\#| ) -> void
def foo; end
```

In combination with the syntax highlighting we added to the Ruby LSP, this makes it easier to read.

The issue is that Rubocop Layout/LeadingCommentSpace will try to add a space before the `|` even with `AllowRBSInlineAnnotation` enabled.

This PR proposes adding `#|` to the list of allowed characters for RBS inline comments.

Signed-off-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>