[Fix #14072] Resolve autocorrect issue in Layout/HashAlignment
This is a fix with two parts.

Part 1 is to remove the attribute max_key_width. The TableAlignment
class, by setting and storing a max_key_width value when processing
the first key-value pair, didn't support hash literals with mixed
separators (colon and hash rocket). The solution is to calculate
the value when needed. This is less efficient, but correct.

Part 2 is to calculate the max width of separators (delimiters) in
the whole hash literal, rather than assuming the width will be the
same for all pairs. This, too, is done to support hashes with mixed
separators.