Fix `Style/PercentQLiterals` cop error on Unicode escape sequence
Prior to these changes this cop would try to take `%q(\u)`, change `q` to `Q`
and parse it. The result (processed source) would contain the `invalid_unicode_escape`
diagnostic error -- but it would be ignored.

Since is that case `ast` is `nil` a `NoMethodError` is thrown.
I don't know if we should only check for this specific diagnostics or all of them.