[Fix #14156] Make `Style/IfUnlessModifier` allow endless method definition
This PR makes `Style/IfUnlessModifier` allow endless method definition in the `if` body.

The code `def method_name = body if condition` is considered a bad case by
`Style/AmbiguousEndlessMethodDefinition` cop. So, to respect the user's intention to use
an endless method definition in the `if` body, the following code is allowed:

```ruby
if condition
  def method_name = body
end
```

Resolves #14156.