Fix `Style/ConditionalAssignment` cop error on `if` in condition
Currently this code leads to an error (when style is set to `assign_in_condition`):

```ruby
if a =
  if b
    1
  else
    2
  end
  42
end
```

since this is a very weird code to write and I don't see a way to consistenly correct it I think it
might make sense not to register an offense.

By the way, it might be a decent idea to detect `if` inside conditions separately.