[Fix #14077] Change `nil` representation in todo file comments
Some cops use `nil` as a supported value for enforced style in their
configuration. This is expressed using tilde (`~`) in the configuration
file. Nil usually means a fallback to some default value.

When we list supported styles in `.rubocop_todo.yml`, we should express
these values as `~`, i.e., using YAML syntax. Before this change the
`nil` values were shown as empty strings, which was confusing.