[Fix #14050] Modify condition for EOL comment
When running with `--disable-uncorrectable`, comments are
automatically added to disable some offenses. If possible, RuboCop
uses end-of-line comments, and otherwise surrounds the offense with
comments before and after. One instance in which an EOL comment
can't be used is when it would end up inside a multiline literal,
e.g., a `%w` array literal.

The condition we used for this case was incorrect. It checked overlap
between the offense range and the literal range, when in fact we
should be checking if the reported line for the offense falls within
the line span of the literal.

We also need to check at an earlier stage than we did, whether the
source line with the added EOL comment would exceed the maximum
line length. This is to not try and add an EOL comment when it's
not possible, and falling back on before/after comments, which
becomes too complicated when there are multiline literals to
deal with.