Add an option to allow custom RuboCop directive modes
This is more of an issue/feature request with a basic implementation proposal.

I am developing a custom RuboCop plugin that performs
certain actions when it encounters comments with a specific format, for example:

```ruby
# rubocop:my-custom-directive-mode foobar
# some code
```

Currently, this triggers an offense from  `Lint/CopDirectiveSyntax`.

I propose adding a new configuration option — either specifically
for `Lint/CopDirectiveSyntax` or globally under `AllCops`.

This option, named for instance `CustomDirectiveModes`,
would default to an empty array (`[]`) and could be overriden
by plugin's configuration.