[Fix #14390] Better handling of `Style/ArgumentsForwarding` for prefix arguments
There is a bit of logic to handle preceeding argument nodes that are not
being forwarded. But the list is not exhaustive
and is missing for example `self` and `const`.

Instead, rely on the already matched arg nodes to use for autocorrect.
`all_anonymous` needs a bit of special handling, because send classification does not contain the first node that
is part of forwarding