[Fix #14413] Fix a false positive for `Layout/EmptyLinesAroundArguments` with multiline strings that contain only whitespace
The previous implementation was doing a bunch of work to decide if a line was
supposed to be considered or not.
For example, Module.new with method definitions may contain newlines that
should be ignored.

This new implementation only considers the ranges between arguments. The old implementationseemed
to have issues with array-style method calls. To be honest, I did not try to understand the original implementation.