Fix flaky spec for `rubocop_extra_features`
It's not safe to assume that it is always nil when the test starts
since other tests could already have accessed it.

I believe it's also not testing much, except that this method is indeed using a instance variable,
so I opted to remove the test.

Writing something that really tests this seems very difficult.

Followup to https://github.com/rubocop/rubocop/pull/14414