[Fix #14435] Fix false negatives for regexp cops when `Lint/DuplicateRegexpCharacterClassElement` is enabled
It does `expressions = expressions.to_a`, which does not create a copy. So other cops
that run after would not see all
the expressions.

`.dup` would solve this, but all this does is work around a bug in `regexp_parser` < 2.7.0.
RuboCop now requires >= 2.9.3 so all that code can simply be removed.

I didn't add a new test because the code that would be tested is gone now.

Also see https://github.com/rubocop/rubocop/pull/13875 for something quite similar.